requirements.txt
beautifulsoup4
coloredlogs
cryptography
feedparser
fuzzywuzzy
geoip2
lxml
netaddr
OTXv2
pillow
Py3DNS
pyOpenSSL
python-Levenshtein
python-magic
requests
requests_toolbelt
shodan
statistics
tldextract
unidecode
validators